pyproject.toml
[tool.poetry]
name = "firedantic-extras"
version = "0.1.0"
description = "Add-on utilities for Firedantic: collection sync, FastAPI pagination, BigQuery schema generation"
authors = ["Lukas Karlsson <karlsson@altissimo.io>"]
license = "BSD-3-Clause"
readme = "README.md"
repository = "https://github.com/altissimo-hq/firedantic-extras"
packages = [
    { include = "firedantic_extras", from = "src" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.10"
firedantic = { git = "https://github.com/altissimo-hq/firedantic.git", branch = "main" }
google-cloud-firestore = ">=2.0"
pydantic = ">=2.0"

[tool.poetry.extras]
fastapi = ["fastapi", "fastapi-pagination"]
bigquery = ["google-cloud-bigquery"]
all = ["fastapi", "fastapi-pagination", "google-cloud-bigquery"]

[tool.poetry.group.fastapi]
optional = true

[tool.poetry.group.fastapi.dependencies]
fastapi = ">=0.100"
fastapi-pagination = ">=0.12"

[tool.poetry.group.bigquery]
optional = true

[tool.poetry.group.bigquery.dependencies]
google-cloud-bigquery = ">=3.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
pytest = "^8.3.0"
pytest-cov = ">=5.0"
ruff = ">=0.9.0"
pre-commit = "^4.5.1"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-m 'not integration'"
markers = [
    "integration: marks tests that require the Firestore emulator (deselect with '-m \"not integration\"')",
]

[tool.skjold]
sources = ["pypa"]

[tool.mypy]
strict = false
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = [
    "fastapi",
    "fastapi.*",
]
ignore_missing_imports = true
